session 1,2, and 3
read blog for description

+++ Coursework/Coursework/cSprite.h
@@ -13,18 +13,18 @@ class cSprite
 {
 private:
 	// position, rotation, scalling
-	glm::ivec2 spritePos2D;
-	glm::ivec2 spriteTranslation;
-	glm::ivec2 spriteCentre;
-	glm::ivec2 spriteScaling;
+	glm::vec2 spritePos2D;
+	glm::vec2 spriteTranslation;
+	glm::vec2 spriteCentre;
+	glm::vec2 spriteScaling;
 	float spriteScalingRotation;
 	float spriteRotation;
 	glm::mat4x4 spriteTransformMatrix;
 	RECT boundingRect;
 
 	//texturing
-	glm::ivec2 spriteVertexData[4];
-	glm::ivec2 spriteTexCoordData[4];
+	glm::vec2 spriteVertexData[4];
+	glm::vec2 spriteTexCoordData[4];
 
 	// Texture width & Height
 	int textureWidth;
@@ -38,8 +38,8 @@ public:
 	~cSprite();			// Destructor
 	void render();
 	void setSpriteTexCoordData();
-	glm::ivec2 getSpritePos();  // Return the sprites current position
-	void setSpritePos(glm::ivec2 sPosition); // set the position of the sprite
+	glm::vec2 getSpritePos();  // Return the sprites current position
+	void setSpritePos(glm::vec2 sPosition); // set the position of the sprite
 	//LPDIRECT3DTEXTURE9 getTexture();  // Return the sprites current image
 	void setTexture(GLuint GLtexID);  // set the image of the sprite
 	void setTextureDimensions(int texWidth, int textHeight);
